long casted on token duration

+++ src/main/java/com/example/pokemonservice/Security/JwtTokenUtil.java
@@ -21,7 +21,7 @@ public class JwtTokenUtil implements Serializable {
     @Serial
     private static final long serialVersionUID = -2550185165626007488L;
 
-    public static final long JWT_TOKEN_VALIDITY = 20 * 60;
+    public static final long JWT_TOKEN_VALIDITY = (long)20 * 60;
 
     @Value("${jwt.secret}")
     private String secret;
